requirements.txt
python-telegram-bot
spotipy
requests
python-dotenv
yt-dlp
yandex-music